Extract ObjectHandler from WorkerTransport
Move the commonobj/obj resolution logic from WorkerTransport.setupMessageHandler
into a reusable ObjectHandler class. This enables sharing the object resolution
logic between the main thread (WorkerTransport) and the renderer worker.

This commit is a part of the renderer-worker series, the worker rendering
stays disabled until the final commit in this series.
